ELEMENTS OF A BEAT Speed and Meter Tempo mark: indicates the speed of the beat, usually expressed in beats per minute (BPM).
Time signature: the stacked numbers dictate the meter of the music. The top number indicates how many beats per measure. The bottom number, the note value of the beat.

Bar line: signifies the end point of a measure/bar. Bar lines work in conjunction with the time signature to create the rhythmic framework of the measure, section or whole piece. Bar lines divide each line into sensibly sized units of rhythm called measures or bars.

Beat positions
Down beat: the first beat of a measure, named after a conductor’s down-swing.

Back beat: the weaker beat(s) in a measure. In popular music, the snare drum is often played strongly on beats 2 and 4. This is commonly referred to as the ‘back beat’.

Up beat: the last beat of a bar, named after a conductor’s up-swing.

Off-beat eighths: eighth notes that fall between the primary beats. Counted as ‘and’ or ‘an’; abbreviated as ‘+’ or ‘&’.

Off-beat sixteenths: sixteenth notes that fall between the eighth notes. Counted as ‘e’ and ‘a’.