I DON’T WANNA LIVE FOREVER – VERSE BEAT
Learn the groove from the verse of I Don't Wanna Live Forever -- taken from where Taylor Swift is singing.

This groove features a half-time, slowed down feel. The entire song is played with a half-time feel, but the verse feels ‘slower’ because the hi-hat is playing 8th notes instead of sixteenth and thirty-second notes.

Half-Time: A half-time groove plays the back-beat on beat 3 (usually on the snare drum) instead of the common back-beat of 2 and 4.
Listen to how much the feel of the groove changes by simply moving the back-beat from beat 3 to the common 2 and 4.